Process-less HTTP client: each job opens a connection to Google API
The previous architecture had a problem when requests multiplexed in a HTTP2 connection timed out (it's not clear if it's due to Google API or the Mint library): the server returned %Mint.HTTPError{module: Mint.HTTP2, reason: :too_many_concurrent_requests}. In other words, as some requests stalled, new ones were created until we reached the max allowed by the server (100 in this case).

While it's obvious that we should reuse connections, this simple architecture now works. If/when we encountered performance issues, it will be the time to devise a pool-based connection, maybe using a higher-level library. other than Mint