Fix incorrect passing of Drive API next page token
It caused a infinite loop where we were loading the same page again and again.

Reason: the args map of the Oban worker was inconsistently constructed with both an atom (:page_token) and a string ("page_token").
The JSON encoder used by Oban retained only one, the string version, which was always set to the first page token.